Converted enginegame to typescript
